# frontend-web-ui/generated/models/CreateExamDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Autogenex API
 * RESTful API for managing courses, tasks (topics), and exams.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateExamDto
 */
export interface CreateExamDto {
    /**
     * 
     * @type {string}
     * @memberof CreateExamDto
     */
    courseId: string;
    /**
     * 
     * @type {number}
     * @memberof CreateExamDto
     */
    points: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateExamDto
     */
    topics?: Array<string>;
}

/**
 * Check if a given object implements the CreateExamDto interface.
 */
export function instanceOfCreateExamDto(value: object): value is CreateExamDto {
    if (!('courseId' in value) || value['courseId'] === undefined) return false;
    if (!('points' in value) || value['points'] === undefined) return false;
    return true;
}

export function CreateExamDtoFromJSON(json: any): CreateExamDto {
    return CreateExamDtoFromJSONTyped(json, false);
}

export function CreateExamDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateExamDto {
    if (json == null) {
        return json;
    }
    return {
        
        'courseId': json['courseId'],
        'points': json['points'],
        'topics': json['topics'] == null ? undefined : json['topics'],
    };
}

export function CreateExamDtoToJSON(json: any): CreateExamDto {
    return CreateExamDtoToJSONTyped(json, false);
}

export function CreateExamDtoToJSONTyped(value?: CreateExamDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'courseId': value['courseId'],
        'points': value['points'],
        'topics': value['topics'],
    };
}